Fix compile time warnings. No functional change, but it allows to build on win64 platforms.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@424266 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native/common/jk_ajp12_worker.c
@@ -313,7 +313,7 @@ static int ajpv12_sendstring(ajp12_endpoint_t * p, const char *buffer)
 {
     int bufferlen;
 
-    if (buffer && (bufferlen = strlen(buffer))) {
+    if (buffer && (bufferlen = (int)strlen(buffer))) {
 #if defined(AS400) || defined(_OSD_POSIX)
         char buf[2048];
         if (bufferlen < 2048) {
